Move test case to live with the cross version tests.
